proper package naming, add Custom View Support, update deps & misc.
