+++ EasyUpiPayment/src/main/java/com/shreyaspatil/easyupipayment/Singleton.kt
@@ -3,7 +3,7 @@ package com.shreyaspatil.easyupipayment
 import com.shreyaspatil.easyupipayment.listener.PaymentStatusListener
 
 internal object Singleton {
-	@set:JvmSynthetic
-	@get:JvmSynthetic
-	internal var listener: PaymentStatusListener? = null
+    @set:JvmSynthetic
+    @get:JvmSynthetic
+    internal var listener: PaymentStatusListener? = null
 }
