+++ EasyUpiPayment/src/main/java/com/shreyaspatil/easyupipayment/listener/PaymentStatusListener.kt
@@ -3,6 +3,6 @@ package com.shreyaspatil.easyupipayment.listener
 import com.shreyaspatil.easyupipayment.model.TransactionDetails
 
 interface PaymentStatusListener {
-	fun onTransactionCompleted(transactionDetails: TransactionDetails)
-	fun onTransactionCancelled()
+    fun onTransactionCompleted(transactionDetails: TransactionDetails)
+    fun onTransactionCancelled()
 }
