+++ EasyUpiPayment/src/main/java/com/shreyaspatil/easyupipayment/model/Payment.kt
@@ -1,0 +1,15 @@
+package com.shreyaspatil.easyupipayment.model
+
+import java.io.Serializable
+
+data class Payment(
+        var currency: String,
+        var vpa: String,
+        var name: String,
+        var payeeMerchantCode: String?,
+        var txnId: String,
+        var txnRefId: String,
+        var description: String,
+        var amount: String,
+        var defaultPackage: String?
+) : Serializable
